FIT AHORA ES DOUBLE

+++ src/lib/caminante/caminante.cpp
@@ -126,7 +126,7 @@ void Caminante::calcMiFit(int **dist, int numCiuds)
 }
 
 //Devuelve el fitness del caminante.
-float Caminante::MyFit()
+double Caminante::MyFit()
 {
     return fitness;
 }
@@ -293,7 +293,7 @@ void Poblacion::calcFit(Caminante &caminate){
 //Devuelve el porcentaje de caminantes que son mejores que el fit que le introducimos,
 //tambien por mejorFit devuelve el fitness del mejor caminante y por media la
 //media de fitness de los caminates
-float Poblacion::stats(float fit,float &mejorFit,float &media){
+double Poblacion::stats(double fit,double &mejorFit,double &media){
     int cont = 0;
     for (int i = 0; i < numCam; i++){
         calcFit(caminantes[i]); // Calcula el fit
@@ -668,7 +668,7 @@ void Poblacion::seleccionar(int n){
                 
                 fit=caminantes[j+(k*i)].MyFit; //Se coge el primero y se compara su fit con el de todos
                 for(int p=(k*i)+1;p<k*(i+1);p++){   //Elige un participante cuyo fit es el mayor de los no elegidos
-                    if(fit<caminantes[p].MyFit&&!elegido[p]){
+                    if(fit<caminantes[p].MyFit && !elegido[p]){
                         fit=caminantes[p].calcMiFit.MyFit;
                         posicion=p;
                     }
